Add exception handling and improve query building in 'getLogoutUrl' method
This commit introduces a check for 'urlEndSession' in order to throw an exception if the URL is missing. It also refactors the code to construct the query string more efficiently and correctly for logout url. Additionally, a 'getResourceOwner' method is added in 'VutOpenIDConnectProvider' and its usage is showcased in 'client.php'.